Move channel traits to the channel module
These traits are a part of the `Channel` API, put them in the `channel` module.
